Add files via upload
Just like a mini play store, we are tried to create a game zone using HTML, CSS and JavaScript. HTML for structure of pages, CSS for styling and JavaScript for validating user's data. User has to enter details, then he/she logged in, then according to category, it will show games and their description and the URL from which he/she can download the game.